feat: Implement item page and refactor, see below
This is a bit chaotic, but here's a list of the most important changes:
When you search in the dictionary you now get a more fleshed out item page, even though this still needs work

Paths have been refactored so that instead of e.g. /dictionary/FR it's now FR/dictionary. This helps with validation of the dynamic route, saving a lot of lines of code in many places

LanguageSelector now also shows that user's native language on the dictionary page so that he can search on both directions.

User's native language and flag are now stored in session, saving calls to the backend