Multiple improvements to navigation and preparation for a new item page, for details see below.
Add new item buttons are in place but as of yet not functional.

Back buttons on the item page lead back to either the list the item was clicked in or the dictionary page, depending on context.

Back buttons on unit details pages don't eat up vertical space, moved to the top center of screen

Adding Lists, Editing items and Adding Dictionary items on the dictionary page is all done with a green button at the bottom right.

On the unit details page, there was no space for a bottom right button, so the add item button appears among already added items.

Minimal styling improvements here and there.

TopButton that were on the item page have been removed, item page is cleaner now.